Add options to sort and overwrite all YML files
The code now presents additional options to sort all YML files in a project, and to overwrite the existing YML file. If there is more than one model and the overwrite flag is not used, a ValueError is thrown. An option to process all files has also been added to _get_models method.